+++ CHANGES.rst
@@ -93,3 +93,4 @@ v1.0.13
 
 v1.0.14[unreleased]
 -----------
+* Remove 'ordered' param in favor of OrderingFilter